Adds CRD and api for Revision (#1)
* Adds CRD and api for Revision

* Code-review: Make cluster weight float64, add .gitattributes for generated files

+++ pkg/apis/picchu/v1alpha1/cluster_types.go
@@ -16,7 +16,7 @@ import (
 type ClusterSpec struct {
 	Enabled bool         `json:"enabled"`
 	Config  *ConfigSpec  `json:"config,omitempty"`
-	Weight  int          `json:"weight"`
+	Weight  float64      `json:"weight"`
 	Account *AccountSpec `json:"account,omitempty"`
 }
